Rename the entry point to index.js
