+++ src/modules/exchange/exchange-api.js
@@ -2,7 +2,7 @@
 
 import { bridgifyObject, onMethod } from 'yaob'
 
-import { type EdgeExchangeCache } from '../../edge-core-index.js'
+import { type EdgeExchangeCache } from '../../index.js'
 import { type ApiInput } from '../root.js'
 import { getExchangeRate } from './exchange-selectors.js'
