+++ src/modules/storage/repo.js
@@ -6,7 +6,7 @@ import {
   type DiskletFolder,
   type EdgeIo,
   type EdgeWalletInfo
-} from '../../edge-core-index.js'
+} from '../../index.js'
 import { sha256 } from '../../util/crypto/crypto.js'
 import { base16, base58, base64 } from '../../util/encoding.js'
 import { RepoFolder } from './repoFolder.js'
