+++ src/util/crypto/crypto.js
@@ -2,7 +2,7 @@
 
 import aesjs from 'aes-js'
 
-import { type EdgeIo } from '../../edge-core-index.js'
+import { type EdgeIo } from '../../index.js'
 import { base16, base64 } from '../encoding.js'
 import { hashjs } from './external.js'
