+++ test/fake-plugins/fake-currency-info.js
@@ -1,6 +1,6 @@
 // @flow
 
-import { type EdgeCurrencyInfo } from '../../src/edge-core-index.js'
+import { type EdgeCurrencyInfo } from '../../src/index.js'
 
 export const fakeCurrencyInfo: EdgeCurrencyInfo = {
   // Basic currency information:
