+++ test/modules/login/lobby.test.js
@@ -4,7 +4,7 @@ import { assert } from 'chai'
 import elliptic from 'elliptic'
 import { describe, it } from 'mocha'
 
-import { makeFakeContexts, makeFakeIos } from '../../../src/edge-core-index.js'
+import { makeFakeContexts, makeFakeIos } from '../../../src/index.js'
 import { type EdgeInternalStuff } from '../../../src/modules/context/internal-api.js'
 import {
   decryptLobbyReply,
